Implement digit string replacement in GetCalibrationValue
A new Dictionary called digitMap has been added to the GetCalibrationValue function to handle the replacement of digit strings with a certain pattern. Also, tests have been updated in the UnitTest1.cs file to reflect this change in the calibration method. This modification is intended to standardize the characters included in the processed lines.